requirements.txt
aiogram==3.23.0
aiosqlite==0.19.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
aiofiles==25.1.0

# Для PostgreSQL (опционально, если используете PostgreSQL вместо SQLite):
# asyncpg==0.29.0

